Negative group IDs

+++ src/sfizz/Defaults.h
@@ -153,7 +153,7 @@ namespace Default
     extern const OpcodeSpec<float> oscillatorModDepth;
     extern const OpcodeSpec<float> oscillatorModDepthMod;
     extern const OpcodeSpec<int32_t> oscillatorQuality;
-    extern const OpcodeSpec<uint32_t> group;
+    extern const OpcodeSpec<int64_t> group;
     extern const OpcodeSpec<float> offTime;
     extern const OpcodeSpec<uint32_t> polyphony;
     extern const OpcodeSpec<uint32_t> notePolyphony;
